pyproject.toml
[tool.poetry]
name = "komtek_test"
version = "0.1.0"
description = ""
authors = ["kiriharu <kiriharu@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
